use app group user defaults for settings
