+++ iOS/Widgets/Widgets.swift
@@ -54,6 +54,11 @@ class Provider: TimelineProvider {
     
     var bookmarksDB: CoreDataDatabase?
 
+    var isAiChatEnabled: Bool {
+        // return (UserDefaults(suiteName: SharedAppUserDefaultsKeys.groupName)?.value(forKey: SharedAppUserDefaultsKeys.isAIChatEnabled) as? Bool) ?? true
+        return false
+    }
+
     func getSnapshot(in context: Context, completion: @escaping (FavoritesEntry) -> Void) {
         createEntry(in: context) { entry in
             completion(entry)
@@ -61,7 +66,7 @@ class Provider: TimelineProvider {
     }
 
     func placeholder(in context: Context) -> FavoritesEntry {
-        return FavoritesEntry(date: Date(), favorites: [], isPreview: context.isPreview)
+        return FavoritesEntry(date: Date(), favorites: [], isPreview: context.isPreview, isAiChatEnabled: isAiChatEnabled)
     }
 
     func getTimeline(in context: Context, completion: @escaping (Timeline<FavoritesEntry>) -> Void) {
@@ -121,11 +126,11 @@ class Provider: TimelineProvider {
             Logger.general.debug("dbFavorites loaded \(dbFavorites.count)")
             let favorites = coreDataFavoritesToFavorites(dbFavorites, returningNoMoreThan: maxFavorites)
             Logger.general.debug("favorites converted \(favorites.count)")
-            let entry = FavoritesEntry(date: Date(), favorites: favorites, isPreview: favorites.isEmpty && context.isPreview)
+            let entry = FavoritesEntry(date: Date(), favorites: favorites, isPreview: favorites.isEmpty && context.isPreview, isAiChatEnabled: isAiChatEnabled)
             Logger.general.debug("entry created")
             completion(entry)
         } else {
-            let entry = FavoritesEntry(date: Date(), favorites: [], isPreview: context.isPreview)
+            let entry = FavoritesEntry(date: Date(), favorites: [], isPreview: context.isPreview, isAiChatEnabled: isAiChatEnabled)
             completion(entry)
         }
     }
@@ -162,6 +167,7 @@ struct FavoritesEntry: TimelineEntry {
     let date: Date
     let favorites: [Favorite]
     let isPreview: Bool
+    let isAiChatEnabled: Bool
 
     func favoriteAt(index: Int) -> Favorite? {
         guard index < favorites.count else { return nil }
